requirements.txt
PyQt6
keyboard
dxcam
mss
sounddevice
webrtcvad
openai-whisper
anthropic
elevenlabs
pyttsx3
pillow
psutil
pywin32
keyring
google-genai
opencv-python
numpy
pytest
pytest-qt
pytest-httpx
